Move processor harness into own module
