+++ crates/lib3h/tests/utils/mod.rs
@@ -2,6 +2,8 @@ pub mod constants;
 #[macro_use]
 pub mod predicate;
 
+pub mod processor_harness;
+
 #[allow(unused_macros)]
 macro_rules! assert_process_success {
     ($node: ident, $req: ident) => {
